feat(all): implement new data protocol
Both sides now use the data packet for comms.
Firmware uses system temperature wrapper for sending temps.
Host uses system data receiver and handler for receiving and parsing.
Data is written to temperature file as xx.xx (rounded).

+++ host/src/app/system_data_handler.c
@@ -5,14 +5,14 @@
 #include "common/protocol.h"
 #include "logic/storage_mgr.h"
 
-static inline float reconstruct_float(uint16_t value);
+static inline float reconstruct_float(int16_t value);
 
 int system_handle_temp(const DataPacket *packet) {
     if (packet->type != TEMP) {
         return -1;
     }
 
-    const uint16_t value = (packet->payload[0] << 8) | packet->payload[1];
+    const int16_t value = (int16_t) (((uint16_t) (packet->payload[0] << 8)) | ((uint16_t) packet->payload[1]));
     const float data = reconstruct_float(value);
     if (storage_mgr_write_temp(data) != 0) {
         return -2;
@@ -21,6 +21,6 @@ int system_handle_temp(const DataPacket *packet) {
     return 0;
 }
 
-static inline float reconstruct_float(uint16_t value) {
-    return value / 100;
+static inline float reconstruct_float(int16_t value) {
+    return value / 100.0f;
 }
